Visual organization of functionality inside the resource panel
