+++ geonode_mapstore_client/client/themes/geonode/less/_resource-details.less
@@ -1,0 +1,31 @@
+// **************
+// Layout
+// **************
+.ms-resource-detail {
+    .gn-details-settings {
+        .permissions-container {
+            max-height: 300px;
+            overflow-y: auto;
+
+            .ms-permissions {
+                width: 100%;
+            }
+            .ms-permissions * {
+                font-size: @font-size-small;
+            }
+        }
+    }
+
+    .gn-details-data {
+        .gn-details-data-table {
+            max-height: 300px;
+            overflow-y: auto;
+        }
+    }
+
+    .gn-details-time-series {
+        .form-group * {
+            font-size: @font-size-small;
+        }
+    }
+}
