+++ geonode_mapstore_client/client/themes/geonode/less/geonode.less
@@ -13,6 +13,7 @@
 @import '_map-viewers-catalog.less';
 @import '_metadata.less';
 @import '_operation.less';
+@import '_resource-details.less';
 @import '_resources-catalog.less';
 @import '_search-bar.less';
 @import '_share.less';
