+++ geonode_mapstore_client/static/mapstore/gn-translations/data.de-DE.json
@@ -198,6 +198,7 @@
             "label": "Beschriftung",
             "description": "Beschreibung",
             "attributes": "Attribute",
+            "data": "Daten",
             "filterBy": "Filtern...",
             "addPermissionsEntry": "Benutzer / Gruppen hinzufügen",
             "groups": "Gruppen",
