+++ geonode_mapstore_client/static/mapstore/gn-translations/data.es-ES.json
@@ -198,6 +198,7 @@
             "label": "Etiqueta",
             "description": "Descripción",
             "attributes": "Atributos",
+            "data": "Datos",
             "filterBy": "Filtro ...",
             "addPermissionsEntry": "Agregar usuarios / grupos",
             "groups": "Grupos",
