+++ geonode_mapstore_client/static/mapstore/gn-translations/data.pt-PT.json
@@ -196,6 +196,7 @@
             "label": "Label",
             "description": "Description",
             "attributes": "Attributes",
+            "data": "Data",
             "filterBy": "Filter...",
             "addPermissionsEntry": "Add Users / Groups",
             "groups": "Groups",
