+++ geonode_mapstore_client/static/mapstore/gn-translations/data.sv-SE.json
@@ -197,6 +197,7 @@
             "label": "Label",
             "description": "Description",
             "attributes": "Attributes",
+            "data": "Data",
             "filterBy": "Filter...",
             "addPermissionsEntry": "Add Users / Groups",
             "groups": "Groups",
